\name{MRRln2pyx}

\alias{MRRln2pyx}

\title{Y on X median rank regression for lognormal 2-parameter } 

\description{
\code{MRRw2pyx} fits a set of data consisting of failures, or alternatively failures and suspensions,
 to the 2-parameter lognormal distribution.
}

\usage{
MRRln2pyx(x,s=NULL, options=NULL)
}

\arguments{
  \item{x}{A vector of failure data.}
  \item{s}{An optional vector of suspension data.}
  \item{options}{An optional list object such as generated by options.abrem including the named item method.fit.
  Implementation only accounts for the second element of method.fit with a value of "qbeta", otherwise the method
  of fitting will default to Benard's approximation for determining plotting positions.}
}

\value{
   A vector containing results in the following order: Mulog, Sigmalog, R_squared
}

\details{
  This function will sort all input data while preserving failure and suspension information in paired vectors.
  Then a call is made to a fast C++ implementation to complete the regression. Y on X regression is considered
  an alternative for possible use with interval censored data by Dr. Abernethy. Refer to section 5.8 of The New
  Weibull Handbook, Fifth Edition.
}

\references{
    Dr. Robert B. Abernethy, (2008) "The New Weibull Handbook, Fifth Edition"
}

\examples{
failures<-c(90,96,30,49,82)
suspensions<-c(100,45,10)
regression_result<-MRRln2pyx(failures,suspensions)
}

\keyword{ regression }