\name{pivotals-package}
\alias{pivotals}
\docType{package}
\title{
Functions to support pivotal analysis studies
}
\description{
Pivotal analysis is used to establish the Critical Correlation Coeficient (CCC) and 
 to establish confidence interval bounds for the fitting of a data set.

}

\section{Introduction}{
	\pkg{pivotals} provides a multi-purpose pivotal engine based on MRR X on Y fitting of random data of the selected model.
}

\author{
David J. Silkworth

Maintainer: David J. Silkworth <davidsilkworth@openreliability.me>
}
\references{
 Dr. Robert B. Abernethy, (2008) "The New Weibull Handbook, Fifth Edition"
}

\keyword{ package }



