\name{pivotals-package}
\alias{pivotals}
\docType{package}
\title{
Functions to support pivotal analysis studies
}
\description{
Pivotal analysis is used to establish goodness of fit by means of Critical Correlation Coeficient (CCC) 
 and the p-value of r and r^2.  This analysis is also used and to establish confidence interval bounds
 for a given data set size and event classification (fail-suspension). 
}
\details{
Functions:
\code{CBpiv}\cr
\code{mrank}\cr
\code{prrVal}\cr


}
\section{Introduction}{
	\pkg{pivotals} provides a multi-purpose pivotal engine based on MRR X on Y fitting of random data for a selected model.
}






\author{
Jacob T. Ormerod

Maintainer: David J. Silkworth <davidsilkworth@openreliability.me>
}
\references{
 Dr. Robert B. Abernethy, (2008) "The New Weibull Handbook, Fifth Edition"
}

\keyword{ package }



